User CRUD added

+++ backend/src/main/java/com/alper/shotify/backend/repository/IUserRepository.java
@@ -3,5 +3,5 @@ package com.alper.shotify.backend.repository;
 import com.alper.shotify.backend.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUserRepository extends JpaRepository<UserEntity, Long> {
+public interface IUserRepository extends JpaRepository<UserEntity, Integer> {
 }
